t0610: skip concurrent write test case on Windows
We tried quite a few things, but this is a failure introduced at the
last -rc before v2.47.0 _and_ it only documents existing behavior as
far as Windows is concerned (concurrent writes are a problem there with
reftables).

So let's punt and simply disable this test for now, to take the pressure
off of v2.47.0.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>